email_spam_detection.py
We’ve all been the recipient of spam emails before. Spam mail, or junk mail, is a type of email
that is sent to a massive number of users at one time, frequently containing cryptic
messages, scams, or most dangerously, phishing content.

In this Project, I used Python to build an email spam detector. Then, used machine learning to train the spam detector to recognize and classify emails into spam and non-spam.